implemented Synology group/useradd modules

+++ salt/modules/useradd.py
@@ -9,6 +9,8 @@ Manage users with the useradd command
     <module-provider-override>`.
 '''
 from __future__ import absolute_import
+import uuid
+import string
 
 try:
     import pwd
@@ -74,9 +76,27 @@ def _build_gecos(gecos_dict):
     and returns a full GECOS comment string, to be used with usermod.
     '''
     return u'{0},{1},{2},{3}'.format(gecos_dict.get('fullname', ''),
-                                    gecos_dict.get('roomnumber', ''),
-                                    gecos_dict.get('workphone', ''),
-                                    gecos_dict.get('homephone', ''))
+                                     gecos_dict.get('roomnumber', ''),
+                                     gecos_dict.get('workphone', ''),
+                                     gecos_dict.get('homephone', ''))
+
+
+def _syno_get_user(name):
+    '''
+    Parse Synology user informations and return them
+    in dictionary form.
+    '''
+    def _c(s):
+        return s.strip('[] ').lower()
+    cmd = ['synouser', '--get', name]
+    user = __salt__['cmd.run_stdout'](cmd, python_shell=False)
+    out = dict(map(_c, s.split(':')) for s in user.split('\n') if ':' in s)
+    groups = [s.split()[1] for s in user.split('\n') if s.startswith('(')]
+    out['expired'] = out['expired'] != 'false'
+    out['member of'] = groups
+    for int_value in ('alloc size', 'primary gid', 'user uid'):
+        out[int_value] = int(out[int_value])
+    return out
 
 
 def _update_gecos(name, key, value, root=None):
@@ -95,10 +115,17 @@ def _update_gecos(name, key, value, root=None):
     gecos_data = copy.deepcopy(pre_info)
     gecos_data[key] = value
 
-    cmd = ['usermod', '-c', _build_gecos(gecos_data), name]
+    if __grains__['os_family'] == 'Synology':
+        user_info = _syno_get_user(name)
+        cmd = ['synouser', '--modify', name,
+               _build_gecos(gecos_data),
+               int(user_info['expired']),
+               user_info['user mail']]
+    else:
+        cmd = ['usermod', '-c', _build_gecos(gecos_data), name]
 
-    if root is not None and __grains__['kernel'] != 'AIX':
-        cmd.extend(('-R', root))
+        if root is not None and __grains__['kernel'] != 'AIX':
+            cmd.extend(('-R', root))
 
     __salt__['cmd.run'](cmd, python_shell=False)
     post_info = info(name)
@@ -130,80 +157,89 @@ def add(name,
 
         salt '*' user.add name <uid> <gid> <groups> <home> <shell>
     '''
-    cmd = ['useradd']
-    if shell:
-        cmd.extend(['-s', shell])
-    if uid not in (None, ''):
-        cmd.extend(['-u', str(uid)])
-    if gid not in (None, ''):
-        cmd.extend(['-g', str(gid)])
-    elif groups is not None and name in groups:
-        defs_file = '/etc/login.defs'
-        if __grains__['kernel'] != 'OpenBSD':
-            try:
-                with salt.utils.files.fopen(defs_file) as fp_:
-                    for line in fp_:
-                        if 'USERGROUPS_ENAB' not in line[:15]:
-                            continue
+    if __grains__['os_family'] == 'Synology':
+        if gid:
+            log.error('specifying user primary group is not supported on this platform')
+            return False
+        cmd = ['synouser', '--add',
+               name, uuid.uuid4(),
+               fullname, '0', 'user@localhost', '0'
+               ]
+    else:
+        cmd = ['useradd']
+        if shell:
+            cmd.extend(['-s', shell])
+        if uid not in (None, ''):
+            cmd.extend(['-u', str(uid)])
+        if gid not in (None, ''):
+            cmd.extend(['-g', str(gid)])
+        elif groups is not None and name in groups:
+            defs_file = '/etc/login.defs'
+            if __grains__['kernel'] != 'OpenBSD':
+                try:
+                    with salt.utils.files.fopen(defs_file) as fp_:
+                        for line in fp_:
+                            if 'USERGROUPS_ENAB' not in line[:15]:
+                                continue
 
-                        if 'yes' in line:
+                            if 'yes' in line:
+                                cmd.extend([
+                                    '-g', str(__salt__['file.group_to_gid'](name))
+                                ])
+
+                            # We found what we wanted, let's break out of the loop
+                            break
+                except OSError:
+                    log.debug(
+                        'Error reading ' + defs_file,
+                        exc_info_on_loglevel=logging.DEBUG
+                    )
+            else:
+                usermgmt_file = '/etc/usermgmt.conf'
+                try:
+                    with salt.utils.files.fopen(usermgmt_file) as fp_:
+                        for line in fp_:
+                            if 'group' not in line[:5]:
+                                continue
+
                             cmd.extend([
-                                '-g', str(__salt__['file.group_to_gid'](name))
+                                '-g', str(line.split()[-1])
                             ])
 
-                        # We found what we wanted, let's break out of the loop
-                        break
-            except OSError:
-                log.debug(
-                    'Error reading ' + defs_file,
-                    exc_info_on_loglevel=logging.DEBUG
-                )
-        else:
-            usermgmt_file = '/etc/usermgmt.conf'
-            try:
-                with salt.utils.files.fopen(usermgmt_file) as fp_:
-                    for line in fp_:
-                        if 'group' not in line[:5]:
-                            continue
+                            # We found what we wanted, let's break out of the loop
+                            break
+                except OSError:
+                    # /etc/usermgmt.conf not present: defaults will be used
+                    pass
 
-                        cmd.extend([
-                            '-g', str(line.split()[-1])
-                        ])
+        if createhome:
+            cmd.append('-m')
+        elif (__grains__['kernel'] != 'NetBSD'
+                and __grains__['kernel'] != 'OpenBSD'):
+            cmd.append('-M')
 
-                        # We found what we wanted, let's break out of the loop
-                        break
-            except OSError:
-                # /etc/usermgmt.conf not present: defaults will be used
-                pass
+        if nologinit:
+            cmd.append('-l')
 
-    if createhome:
-        cmd.append('-m')
-    elif (__grains__['kernel'] != 'NetBSD'
-            and __grains__['kernel'] != 'OpenBSD'):
-        cmd.append('-M')
+        if home is not None:
+            cmd.extend(['-d', home])
 
-    if nologinit:
-        cmd.append('-l')
+        if not unique and __grains__['kernel'] != 'AIX':
+            cmd.append('-o')
 
-    if home is not None:
-        cmd.extend(['-d', home])
+        if (system
+            and __grains__['kernel'] != 'NetBSD'
+                and __grains__['kernel'] != 'OpenBSD'):
+            cmd.append('-r')
 
-    if not unique and __grains__['kernel'] != 'AIX':
-        cmd.append('-o')
+        if __grains__['kernel'] == 'OpenBSD':
+            if loginclass is not None:
+                cmd.extend(['-L', loginclass])
 
-    if (system
-        and __grains__['kernel'] != 'NetBSD'
-        and __grains__['kernel'] != 'OpenBSD'):
-        cmd.append('-r')
+        cmd.append(name)
 
-    if __grains__['kernel'] == 'OpenBSD':
-        if loginclass is not None:
-            cmd.extend(['-L', loginclass])
-
-    cmd.append(name)
-
-    if root is not None and __grains__['kernel'] != 'AIX':
-        cmd.extend(('-R', root))
+        if root is not None and __grains__['kernel'] != 'AIX':
+            cmd.extend(('-R', root))
 
     ret = __salt__['cmd.run_all'](cmd, python_shell=False)
 
@@ -241,18 +277,30 @@ def delete(name, remove=False, force=False, root=None):
 
         salt '*' user.delete name remove=True force=True
     '''
-    cmd = ['userdel']
+    if __grains__['os_family'] == 'Synology':
+        if remove:
+            cmd = ['synouser', '--del', name]
+        else:
+            # XXX: rename user ?
+            user_info = _syno_get_user(name)
+            gecos_data = _get_gecos(name)
+            cmd = ['synouser', '--modify', name,
+                   _build_gecos(gecos_data),
+                   1,
+                   user_info['user mail']]
+    else:
+        cmd = ['userdel']
 
-    if remove:
-        cmd.append('-r')
+        if remove:
+            cmd.append('-r')
 
-    if force and __grains__['kernel'] != 'OpenBSD' and __grains__['kernel'] != 'AIX':
-        cmd.append('-f')
+        if force and __grains__['kernel'] != 'OpenBSD' and __grains__['kernel'] != 'AIX':
+            cmd.append('-f')
 
-    cmd.append(name)
+        cmd.append(name)
 
-    if root is not None and __grains__['kernel'] != 'AIX':
-        cmd.extend(('-R', root))
+        if root is not None and __grains__['kernel'] != 'AIX':
+            cmd.extend(('-R', root))
 
     ret = __salt__['cmd.run_all'](cmd, python_shell=False)
 
@@ -311,6 +359,11 @@ def chuid(name, uid):
     pre_info = info(name)
     if uid == pre_info['uid']:
         return True
+
+    if __grains__['os_family'] == 'Synology':
+        log.error('user.chuid is not yet supported on this platform')
+        return False
+
     cmd = ['usermod', '-u', '{0}'.format(uid), name]
     __salt__['cmd.run'](cmd, python_shell=False)
     return info(name).get('uid') == uid
@@ -329,6 +382,11 @@ def chgid(name, gid, root=None):
     pre_info = info(name)
     if gid == pre_info['gid']:
         return True
+
+    if __grains__['os_family'] == 'Synology':
+        log.error('user.chgid is not yet supported on this platform')
+        return False
+
     cmd = ['usermod', '-g', '{0}'.format(gid), name]
 
     if root is not None and __grains__['kernel'] != 'AIX':
@@ -351,6 +409,11 @@ def chshell(name, shell, root=None):
     pre_info = info(name)
     if shell == pre_info['shell']:
         return True
+
+    if __grains__['os_family'] == 'Synology':
+        log.error('user.chshell is not yet supported on this platform')
+        return False
+
     cmd = ['usermod', '-s', shell, name]
 
     if root is not None and __grains__['kernel'] != 'AIX':
@@ -374,6 +437,11 @@ def chhome(name, home, persist=False, root=None):
     pre_info = info(name)
     if home == pre_info['home']:
         return True
+
+    if __grains__['os_family'] == 'Synology':
+        log.error('user.chhome is not yet supported on this platform')
+        return False
+
     cmd = ['usermod', '-d', '{0}'.format(home)]
 
     if root is not None and __grains__['kernel'] != 'AIX':
@@ -384,6 +452,26 @@ def chhome(name, home, persist=False, root=None):
     cmd.append(name)
     __salt__['cmd.run'](cmd, python_shell=False)
     return info(name).get('home') == home
+
+
+def _syno_manage_group(action, group, name):
+    '''
+    Modify group membership on a Synology NAS
+
+    :param action: either `remove` or `add`
+    :param group: group name we are changing
+    :param name: username to add/remove
+    '''
+    cmd = ['synogroup', '--get', group]
+    ret = __salt__['cmd.run_stdout'](cmd, python_shell=False)
+    group_users = set(x.split(':')[1].strip('[]')
+                      for x
+                      in ret.split('\n')
+                      if x
+                      and x[0] in string.digits)
+    getattr(group_users, action)(name)
+    cmd = ['synogroup', '--member', group, ' '.join(group_users)]
+    return __salt__['cmd.retcode'](cmd, python_shell=False)
 
 
 def chgroups(name, groups, append=False, root=None):
@@ -412,6 +500,21 @@ def chgroups(name, groups, append=False, root=None):
     ugrps = set(list_groups(name))
     if ugrps == set(groups):
         return True
+
+    if __grains__['os_family'] == 'Synology':
+
+        rets = []
+        groups = set(groups)
+
+        if not append:
+            for group in ugrps - groups:
+                rets += _syno_manage_group('remove', group, name)
+
+        for group in groups - ugrps:
+            rets += _syno_manage_group('add', group, name)
+
+        return not any(rets)
+
     cmd = ['usermod']
 
     if __grains__['kernel'] != 'OpenBSD':
